fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,9 +5,9 @@
  */
 int main(void)
 {
-	int f1 = 1;
-	int f2 = 1;
-	int fib = 0;
+	long int f1 = 1;
+	long int f2 = 1;
+	long int fib = 0;
 	int n;
 
 	printf("%d, %d, ", f1, f2);
